requirements.txt
openai-whisper
torch
numpy
